+++ pyaud/_main.py
@@ -27,7 +27,7 @@ def main() -> None:
     _plugins.load()
     parser = _Parser(_colors.cyan.get(_environ.NAME))
     _initialize_dirs()
-    _data.read(_data.record, _environ.DATADIR / _data.DURATIONS)
+    _data.read(_data.record, _environ.DURATIONS_FILE)
     _config.load_config(parser.args.rcfile)
     _config.configure_logging(parser.args.verbose)
     _files.add_exclusions(*_config.toml["indexing"]["exclude"])