+++ pyaud/parsers.py
@@ -10,6 +10,8 @@ from pathlib import Path as _Path
 
 import m2r as _m2r
 
+from ._environ import environ as _environ
+
 
 class Md2Rst:
     """Convert markdown file into RST file.
@@ -49,13 +51,13 @@ class LineSwitch:
 
     def __init__(self, path: _Path, obj: _t.Dict[int, str]) -> None:
         self._path = path
-        self.read = path.read_text(encoding="utf-8")
+        self.read = path.read_text(encoding=_environ.ENCODING)
         edit = self.read.splitlines()
         for count, _ in enumerate(edit):
             if count in obj:
                 edit[count] = obj[count]
 
-        path.write_text("\n".join(edit), encoding="utf-8")
+        path.write_text("\n".join(edit), encoding=_environ.ENCODING)
 
     def __enter__(self) -> LineSwitch:
         return self
@@ -63,4 +65,4 @@ class LineSwitch:
     def __exit__(
         self, exc_type: _t.Any, exc_val: _t.Any, exc_tb: _t.Any
     ) -> None:
-        self._path.write_text(self.read, encoding="utf-8")
+        self._path.write_text(self.read, encoding=_environ.ENCODING)